requirements.txt
Django==1.8.3
SQLAlchemy>=0.7.3
alembic==0.6.2
djangorestframework>=3.4.0
drf-nested-routers>=0.11.1
iso8601>=0.1.10
pytz>=2016.6.1
python-memcached==1.53
psycopg2>=2.6.1
pyyaml>=3.11
bcrypt>=2.0.0
oslo.db>=4.8.0

